backend/graph_engine/01_Seed_Base_Taxonomy/requirements.txt
neo4j>=5.0.0
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
gitpython>=3.1.40
python-Levenshtein>=0.23.0
